Fix progress bar bug after resuming fetch.
It reached 100% too early, as the maximum value was set to the number of tracks
to fetch info for, while the current value was set to the current index in the list.

That means that if you abort the fetch at track 150/400, the maximum will then start at 250,
but the value will start at 150. Thus it will display 100% after doing only 100 more tracks,
for a total of 250/400 tracks in this example, and then remain at 100% until actually finishing.